+++ src/event/event.rs
@@ -25,25 +25,25 @@ use piston_window::*;
 pub type PushrodEventMask = u32;
 
 /// No events mask.
-pub const PUSHROD_EVENT_NONE: PushrodEventMask = 0x00000000;
+pub const MASK_EVENT_NONE: PushrodEventMask = 0x00000000;
 
 /// Mouse movement event mask.
-pub const PUSHROD_EVENT_MOUSE_MOVED: PushrodEventMask = 0x00000001;
+pub const MASK_EVENT_MOUSE_MOVED: PushrodEventMask = 0x00000001;
 
 /// Mouse button press event mask.
-pub const PUSHROD_EVENT_MOUSE_DOWN: PushrodEventMask = 0x00000002;
+pub const MASK_EVENT_MOUSE_DOWN: PushrodEventMask = 0x00000002;
 
 /// Mouse button release event mask.
-pub const PUSHROD_EVENT_MOUSE_UP: PushrodEventMask = 0x00000004;
+pub const MASK_EVENT_MOUSE_UP: PushrodEventMask = 0x00000004;
 
 /// Mouse scroll event mask.
-pub const PUSHROD_EVENT_MOUSE_SCROLL: PushrodEventMask = 0x00000008;
+pub const MASK_EVENT_MOUSE_SCROLL: PushrodEventMask = 0x00000008;
 
 /// All mouse events mask.
-pub const PUSHROD_EVENT_MOUSE_ALL: PushrodEventMask = 0x0000000F;
+pub const MASK_EVENT_MOUSE_ALL: PushrodEventMask = 0x0000000F;
 
 /// All events mask.  (Use this carefully)
-pub const PUSHROD_EVENT_ALL: PushrodEventMask = PUSHROD_EVENT_MOUSE_ALL;
+pub const MASK_EVENT_ALL: PushrodEventMask = MASK_EVENT_MOUSE_ALL;
 
 /// Enumeration types for different `PushrodEvents`, translated from the underlying OS.
 pub enum PushrodEvent {
@@ -80,7 +80,7 @@ pub enum PushrodEvent {
 /// #
 /// impl PushrodEventListener for EventListener {
 ///     fn event_mask(&self) -> PushrodEventMask {
-///         PUSHROD_EVENT_MOUSE_MOVED
+///         MASK_EVENT_MOUSE_MOVED
 ///     }
 ///
 ///     fn handle_event(&self, event: &PushrodEvent) {
@@ -94,12 +94,12 @@ pub enum PushrodEvent {
 ///
 /// Programmers who use this event system are encouraged to override `event_mask` so that
 /// they only receive the events that pertain to your application.  If this is strictly set
-/// to `PUSHROD_EVENT_ALL`, all wrapped events will be sent to `handle_event`.
+/// to `MASK_EVENT_ALL`, all wrapped events will be sent to `handle_event`.
 pub trait PushrodEventListener {
     /// Identifies which events to receive in the `handle_event` function.  Any events that
     /// do not match the masks (defined in constants) will not trigger a `handle_event` callback.
     fn event_mask(&self) -> PushrodEventMask {
-        PUSHROD_EVENT_ALL
+        MASK_EVENT_ALL
     }
 
     /// Called when an event matching a masked type (in `event_mask`) occurs.